+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JW6E4504TWWK1JK02WR77E2Q"
+        "01JW83C6TJWMYRFV3790C8ZG29"
     )
     r"""The id of the resource"""