+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -59,5 +59,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-05-26T13:57:08.249Z")
+    created: Optional[datetime] = parse_datetime("2025-05-27T05:27:46.805Z")
     r"""The date and time the resource was created"""