+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JW6E4502DDS1QM7MY4NSBR7S"
+        "01JW83C6THBFFJNGKQNQH8TBBE"
     )
     r"""The id of the resource"""