+++ packages/orq-rc/RELEASES.md
@@ -1529,3 +1529,13 @@ Based on:
 - [python v3.6.0-rc.38] packages/orq-rc
 ### Releases
 - [PyPI v3.6.0-rc.38] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.38 - packages/orq-rc
+
+## 2025-05-27 05:29:10
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.551.0 (2.610.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.6.0-rc.39] packages/orq-rc
+### Releases
+- [PyPI v3.6.0-rc.39] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.39 - packages/orq-rc